extracted decideSegmentListRangeForTemplate method to SegmentsUtil
